requirements.txt
asgiref==3.5.2
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.2
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-filter==22.1
django-js-asset==2.0.0
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
environs==9.5.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
phonenumbers==8.13.0
Pillow==9.3.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
rfc3986==1.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sniffio==1.3.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
